Time: 3 ms (95.29%), Space: 10.3 MB (86.34%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int change(int amount, vector<int>& coins) {
-        vector<unsigned long long> dp(amount+1, 0);
+        vector<double> dp(amount+1, 0);
         dp[0]=1;
         for(long long i=0;i<coins.size();i++){
             for(long long j=coins[i];j<=amount;j++){
